+++ praetorian_cli/sdk/entities/attributes.py
@@ -17,10 +17,10 @@ class Attributes:
         """ Delete an attribute """
         return self.api.delete_by_key('attribute', key)
 
-    def list(self, prefix_filter='', source_key=None, offset=None, pages=10000):
+    def list(self, prefix_filter='', source_key=None, offset=None, pages=100000) -> tuple:
         """ List attribute, optionally prefix-filtered by the portion of the key after
             '#attribute#' """
         if source_key:
-            return self.api.search.by_source(source_key, offset, pages)
+            return self.api.search.by_source(source_key, 'attribute', offset, pages)
         else:
             return self.api.search.by_key_prefix(f'#attribute#{prefix_filter}', offset, pages)