+++ praetorian_cli/sdk/test/test_asset.py
@@ -1,6 +1,6 @@
 import pytest
 
-from praetorian_cli.sdk.model.globals import Asset
+from praetorian_cli.sdk.model.globals import Asset, Kind
 from praetorian_cli.sdk.test.utils import make_test_values, clean_test_entities, setup_chariot
 
 
@@ -24,15 +24,7 @@ class TestAsset:
     def test_list_asset(self):
         results, _ = self.sdk.assets.list()
         assert len(results) > 0
-        assert any(a['dns'] == self.asset_dns for a in results)
-
-    def test_jobs_kicked_off(self):
-        jobs, offset = self.sdk.jobs.list(self.asset_dns)
-        assert jobs is not None
-        assert jobs is not []
-        for job in jobs:
-            assert job['source'] is not ''
-            assert job['status'] is not None
+        assert any([a['dns'] == self.asset_dns for a in results])
 
     def test_update_asset(self):
         self.sdk.assets.update(self.asset_key, Asset.FROZEN.value)
@@ -40,7 +32,9 @@ class TestAsset:
 
     def test_delete_asset(self):
         self.sdk.assets.delete(self.asset_key)
-        assert self.sdk.assets.get(self.asset_key)['status'] == Asset.DELETED.value
+        assert self.get_asset() == None
+        deleted_assets, _ = self.sdk.search.by_status(Asset.DELETED.value, Kind.ASSET.value)
+        assert any([a['dns'] == self.asset_dns for a in deleted_assets])
 
     def get_asset(self):
         return self.sdk.assets.get(self.asset_key)