+++ praetorian_cli/sdk/test/test_risk.py
@@ -1,6 +1,6 @@
 import pytest
 
-from praetorian_cli.sdk.model.globals import Risk
+from praetorian_cli.sdk.model.globals import Risk, Kind
 from praetorian_cli.sdk.test.utils import make_test_values, clean_test_entities, setup_chariot
 
 
@@ -25,7 +25,15 @@ class TestRisk:
     def test_list_risks(self):
         results, _ = self.sdk.risks.list()
         assert len(results) > 0
-        assert any(r['dns'] == self.asset_dns for r in results)
+        assert any([r['dns'] == self.asset_dns for r in results])
+
+    def test_bidirectional_links(self):
+        affected_assets = self.sdk.risks.affected_assets(self.risk_key)
+        assert len(affected_assets) == 1
+        assert affected_assets[0]['key'] == self.asset_key
+        associated_risks = self.sdk.assets.associated_risks(self.asset_key)
+        assert len(associated_risks) == 1
+        assert associated_risks[0]['key'] == self.risk_key
 
     def test_update_risk(self):
         self.sdk.risks.update(self.risk_key, Risk.OPEN_CRITICAL.value)
@@ -33,7 +41,9 @@ class TestRisk:
 
     def test_delete_risk(self):
         self.sdk.risks.delete(self.risk_key, Risk.DELETED_DUPLICATE_CRITICAL.value)
-        assert self.get_risk()['status'] == Risk.DELETED_DUPLICATE_CRITICAL.value
+        assert self.get_risk() is None
+        deleted_risks, _ = self.sdk.search.by_status(Risk.DELETED_DUPLICATE_CRITICAL.value, Kind.RISK.value)
+        assert any([r['name'] == self.risk_name for r in deleted_risks])
 
     def get_risk(self):
         return self.sdk.risks.get(self.risk_key)