+++ praetorian_cli/sdk/test/test_z_cli.py
@@ -34,7 +34,7 @@ class TestZCli:
         self.verify(f'get asset "{o.asset_key}"', [o.asset_key, f'"status": "{Asset.FROZEN.value}"'])
 
         self.verify(f'delete asset "{o.asset_key}"')
-        self.verify(f'get asset "{o.asset_key}"', [f'"status": "{Asset.DELETED.value}"'])
+        self.verify(f'get asset "{o.asset_key}"')
 
         clean_test_entities(self.sdk, o)
 
@@ -88,7 +88,7 @@ class TestZCli:
         self.verify(f'get risk "{o.risk_key}"', [o.risk_key, f'"status": "{Risk.OPEN_LOW.value}"'])
 
         self.verify(f'delete risk "{o.risk_key}"')
-        self.verify(f'get risk "{o.risk_key}"', [f'"status": "{Risk.DELETED_OTHER_LOW.value}"'])
+        self.verify(f'get risk "{o.risk_key}"')
 
         clean_test_entities(self.sdk, o)
 
@@ -163,20 +163,20 @@ class TestZCli:
         self.verify(f'search -t "#asset#{o.asset_dns}" -d -p all', [o.asset_key, '"key"', '"data"'])
         self.verify(f'search -t "#asset#{o.asset_dns}" -c -p all', ['"A": 1'])
 
-        self.verify(f'search -t "source:{o.asset_key}" -p all', ['surface#provided', o.asset_key, 'attribute'])
-        self.verify(f'search -t "ip:{o.asset_name}" -p all', [o.asset_key])
-        self.verify(f'search -t "name:{o.asset_name}" -p all', [o.asset_key])
-        self.verify(f'search -t "dns:{o.asset_dns}" -p all', [o.asset_key])
+        self.verify(f'search -t "source:{o.asset_key}" -k attribute -p all',
+                    ['surface#provided', o.asset_key, 'attribute'])
+        self.verify(f'search -t "name:{o.asset_name}" -k asset -p all', [o.asset_key])
+        self.verify(f'search -t "dns:{o.asset_dns}" -k asset -p all', [o.asset_key])
 
-        self.verify(f'search -t "source:{o.asset_key}" -p all', ['surface#provided', o.asset_key, 'attribute'])
-        self.verify(f'search -t "ip:{o.asset_name}" -p all', [o.asset_key])
-        self.verify(f'search -t "name:{o.asset_name}" -p all', [o.asset_key])
-        self.verify(f'search -t "dns:{o.asset_dns}" -p all', [o.asset_key])
-        self.verify(f'search -t "status:{Asset.ACTIVE.value}" -p all', [o.asset_key])
+        self.verify(f'search -t "source:{o.asset_key}" -k attribute -p all',
+                    ['surface#provided', o.asset_key, 'attribute'])
+        self.verify(f'search -t "name:{o.asset_name}" -k asset -p all', [o.asset_key])
+        self.verify(f'search -t "dns:{o.asset_dns}" -k asset -p all', [o.asset_key])
+        self.verify(f'search -t "status:{Asset.ACTIVE.value}" -k asset -p all', [o.asset_key])
 
         self.verify(f'add attribute -n {o.attribute_name} -v {o.attribute_value} -k "{o.asset_key}"')
 
-        self.verify(f'search -t "name:{o.attribute_name}" -p all', [o.asset_key, 'attribute'])
+        self.verify(f'search -t "name:{o.attribute_name}" -k attribute -p all', [o.asset_key, 'attribute'])
 
         clean_test_entities(self.sdk, o)